cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stoa-cli"
version = "0.1.0"
description = "STOA Platform CLI — manage APIs, tenants, and MCP tools"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
]

[project.scripts]
stoa = "src.main:app_runner"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "black>=24.0",
    "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM", "S", "DTZ", "LOG", "RUF"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "ARG001", "ARG002"]
"src/config.py" = ["S105"]
"src/commands/*.py" = ["B008", "S105"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-fail-under=70"
